fix: cast m_posBits to unsigned int in AvailableBits() to avoid -Werror=sign-compare
m_posBits is int, comparing with static_cast<unsigned int>(length) * 8
triggers -Werror=sign-compare in GCC. Cast m_posBits to unsigned int
since it's always >= 0 (bit position).

+++ xbmc/utils/BitstreamReader.h
@@ -23,7 +23,7 @@ public:
   // huge positive count that keeps callers looping forever.
   unsigned int AvailableBits()
   {
-    return m_posBits >= static_cast<unsigned int>(length) * 8 ? 0 : length * 8 - m_posBits;
+    return static_cast<unsigned int>(m_posBits) >= static_cast<unsigned int>(length) * 8 ? 0 : length * 8 - m_posBits;
   }
 
 private:
